Safari-609.1.11.1/LayoutTests/accessibility/gtk/caret-browsing-text-focus-expected.txt

A paragraph with a
line break in the middle

A paragraph with a single line

This tests the FOCUSABLE and FOCUSED states for text objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axPar1.role is 'AXRole: AXParagraph'
PASS axPar2.role is 'AXRole: AXParagraph'
PASS axPar1.isFocusable is true
PASS axPar2.isFocusable is true
PASS axPar1.isFocused is false
PASS axPar2.isFocused is false
PASS axPar1.isFocused is true
PASS axPar2.isFocused is false
PASS axPar1.isFocused is true
PASS axPar2.isFocused is false
PASS axPar1.isFocused is false
PASS axPar2.isFocused is true
PASS axPar1.isFocused is true
PASS axPar2.isFocused is false
PASS axPar1.isFocused is false
PASS axPar2.isFocused is true
PASS successfullyParsed is true

TEST COMPLETE
